Adding LDAP groups support to 'posix-search' LDAP auth
This patch adds basic groups support to POSIX Search-based LDAP authorization.
If a LDAP group DN is set on the settings page - before binding as a user,
TeamPass will do another LDAP lookup to check if a user is a member of specified
group. If it's not - LDAP authorization will fail.
This is useful if you want to limit users of some particular group to be able to
log in.