Sanitize ampersand to URL encoding  in csrfp.config
The verifyGetFor array element in
includes\libraries\csrfp\libs\csrfp.config.sample.php contains an
ampersand as part of a URL. Inserted directly into HTML, it looks like
this:

<input type="hidden" id="csrfp_hidden_data_urls"
value='["*page=items&type=duo_check*"]' />

The ampersand is parsed as the start of an HTML entity (like &nbsp;).
and breaks interpretation of the rest of the HTML element.

This quick fix replaces the & character with %26 which is the ampersand
url-encoded. Hopefully this doesn't break the purpose of this code on a
particular Duo Security page I cannot see.