Folder and item edit dialog box language fixes
1. "Managed by" "Managers of role" in edit users only had one quote in
front of role, removed the unclosed quote so there is no quote around
the role label.
2. Colons had an unneccesary space in front of them. Actually the colons
are all unneccessary when used as labels, removed colons from the labels
in add/edit folders and add/edit items. This should improve language
compatibility in RTL languages.
3. Removed remaining quotes around role names in other edit boxes. Other
languages may not use quotes in this way.
4. "Create folder and role for" domain option in add user is now hidden
until an email domain is input.