New dockerfile and docker-compose.yml, docs update
This approach uses Nginx and more up to date dependencies, as well as converting the images to Alpine. A new Docker Compose example with SSL support is provided, and the docs have been updated on how to use them. Also declutters the repository a little bit by removing apache.conf files