Update export.queries.php
Added the "download" attribute. This attribute allows to download the file referenced in the "href" attribute when clicking on the link.

In my opinion, as it was, the process of obtaining a file that could be used offline was not transparent... For less tech savvy users, it could lead to some problems.