Update Dockerfile to use $uri and $args variables
Since the Docker-Containers were having trouble with the API, the provided solution (adding the /api/ location to NGINX) solves the problem, however it was implemented wrong. Those variables have to be escaped in the Dockerfile, otherwise the value of the variable will be added (in this case a empty value).