Removed base64_decode used against plaintext in downloadFile.php
This was resulting in either garbled filenames when trying to download attachments, or causing the server to outright refuse the download due URL-unfriendly characters.  It seems self-evident by the use of the basename() that the 'name' and 'extension' fields are stored in plaintext in the {prefix}_files table, so I can only assume this was a oversight and that it's still in the code since 3.0.0 because few people use the attachments feature.